jobber.py/requirements.txt
Django==2.2.6
psycopg2==2.8.3
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
djangorestframework-simplejwt==4.3.0
drf-flex-fields==0.6.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-elasticsearch-dsl==7.1.0
factory-boy==2.12.0
Pillow==6.2.0
tqdm==4.36.1